pyproject.toml
[tool.poetry]
name = "doctree"
version = "0.1.0"
description = "Directory Structure Documentation Generator and Enforcer"
authors = ["Tim Camise <tcamise@gopro.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/tcamise-gpsw/doctree"
documentation = "https://github.com/tcamise-gpsw/doctree"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]

[tool.poetry.dependencies]
python = "^3.10"
rich = "^13.3.4"
pyparsing = "^3.0.9"
gitpython = "^3.1.31"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.1"
pydocstyle = {extras = ["toml"], version = "^6.3.0"}
pytest-cov = "^4.0.0"
pytest-html = "^3.2.0"
coverage = "^7.2.3"
pylint = "^2.17.2"
mypy = "^1.2.0"
sphinx = "^6.1.3"
sphinx-rtd-theme = "^1.2.0"
sphinxemoji = "^0.2.0"
coverage-badge = "1.1.0"
darglint = "^1.8.1"
poethepoet = "^0.19.0"
nox = "^2022.11.21"
nox-poetry = "^1.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
doctree-gen = "doctree.scripts.main:entrypoint"

[tool.poe.tasks.tests]
cmd = "pytest test --cov-fail-under=90"
help = "Run unit tests"

[tool.poe.tasks.types]
cmd = "mypy doctree"
help = "Check types"

[tool.poe.tasks.lint]
cmd = "pylint --no-docstring-rgx=__|main|parse_arguments|entrypoint doctree"
help = "Run pylint"

[tool.poe.tasks.format]
cmd = "black doctree test"
help = "Apply black formatting to source code"

[tool.poe.tasks.pydocstyle]
cmd = "pydocstyle --config pyproject.toml -v doctree"
help = "check docstrings style"

[tool.poe.tasks.darglint]
cmd = "darglint -v 2 doctree"
help = "validate docstrings"

[tool.poe.tasks.docstrings]
sequence = ["pydocstyle", "darglint"]
help = "Format, check types, lint, and check docstrings"

[tool.poe.tasks.sphinx]
cmd = "sphinx-build -W --keep-going -a -E -b html docs docs/build"
help = "Build sphinx documentation."

[tool.poe.tasks.coverage]
cmd = "coverage-badge -f -o docs/_static/coverage.svg"
help = "update coverage badge"

[tool.poe.tasks.protobuf]
cmd = "bash ./tools/build_protos.sh"
help = "generate protobuf source from .proto (assumes protoc >= 3.20.1 available)"

[tool.poe.tasks.clean_artifacts]
cmd = "rm -rf **/__pycache__ .mypy_cache .nox"
help = "Clean testing artifacts and pycache"

[tool.poe.tasks.clean_tests]
cmd = "rm -rf .reports && rm -rf .pytest_cache"
help = "Clean test reports"

[tool.poe.tasks.clean_docs]
cmd = "rm -f docs/modules.rst && rm -rf docs/build"
help = "Clean built docs output"

[tool.poe.tasks.clean_build]
cmd = "rm -rf dist"
help = "Clean module build output"

[tool.poe.tasks.docs]
sequence = ["docstrings", "sphinx"]
help = "Validate docstrings and build docs"

[tool.poe.tasks.clean]
sequence = ["clean_artifacts", "clean_tests", "clean_docs", "clean_build"]
help = "Clean everything"

[tool.poe.tasks.all]
sequence = ["format", "types", "lint", "tests", "docs"]
help = "Format, check types, lint, check docstrings, and run unit tests"

[tool.mypy]
ignore_missing_imports = true
warn_redundant_casts = true
disallow_untyped_calls = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = false
warn_unused_ignores = true
warn_unreachable = false
show_error_context = true
pretty = true

[tool.pytest.ini_options]
log_auto_indent = true
log_cli = true
log_cli_level = "ERROR"
log_cli_format = "%(asctime)s %(levelname)s %(message)s"
log_cli_date_format = "%H:%M:%S"
log_file_level = "DEBUG"
log_file_format = "%(threadName)13s: %(name)40s:%(lineno)5d %(asctime)s.%(msecs)03d %(levelname)-8s | %(message)s"
log_file_date_format = "%H:%M:%S"
filterwarnings = "ignore::DeprecationWarning"
norecursedirs = "vectors"
addopts = [
    "-s",
    "--capture=tee-sys",
    "--cov=doctree",
    "--cov-report=term",
    "--cov-report=html",
    "--html=.reports/test_report.html",
    "--self-contained-html"
]

[tool.coverage.run]
data_file = ".reports/coverage/.coverage"
branch = true
source = ["doctree"]
omit = ["doctree/main.py"]

[tool.coverage.html]
directory = ".reports/coverage"

[tool.coverage.report]
exclude_lines = ["raise NotImplementedError"]

[tool.pylint.'MASTER']
load-plugins = "pylint.extensions.docparams"
accept-no-param-doc = "yes"
accept-no-return-doc = "yes"
default-docstring-type = "google"
ignore = ["test"]

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "unnecessary-lambda",
    "unnecessary-lambda-assignment",
    "too-many-ancestors",
    "no-name-in-module",
    "too-many-nested-blocks",
    "unspecified-encoding",
    "consider-using-with",
    "abstract-method",
    "useless-type-doc",
    "cyclic-import",
    "logging-fstring-interpolation",
    "logging-format-interpolation",
    "duplicate-code",
    "global-statement",
    "too-few-public-methods",
    "too-many-public-methods",
    "too-many-arguments",
    "too-many-instance-attributes",
    "too-many-branches",
    "too-many-locals",
    "too-many-lines",
    "too-many-statements",
    "fixme",
    "protected-access",
    "invalid-name",
    "unsubscriptable-object",

]

[tool.pylint.'FORMAT']
max-line-length = 160 # This is instead handled by black

[tool.black]
line-length = 111
exclude = ".venv"

[tool.pydocstyle]
convention = "google"
add-ignore = "D415, D107, D105"
match-dir = '(?!.*scripts).*'